refactor and add e2e for h1
